requirement.txt
flask
flask-sqlalchemy
werkzeug
google-cloud-dialogflow
scikit-learn
pandas
joblib